+++ language/src/main/java/i4gl/nodes/builtin/LengthBuiltinNode.java
@@ -6,14 +6,8 @@ import com.oracle.truffle.api.nodes.NodeInfo;
 import i4gl.exceptions.NotImplementedException;
 import i4gl.runtime.types.BaseType;
 import i4gl.runtime.types.primitive.IntType;
-import i4gl.runtime.values.BigIntArray;
+import i4gl.runtime.values.Array;
 import i4gl.runtime.values.Char;
-import i4gl.runtime.values.CharArray;
-import i4gl.runtime.values.FloatArray;
-import i4gl.runtime.values.IntArray;
-import i4gl.runtime.values.Null;
-import i4gl.runtime.values.SmallFloatArray;
-import i4gl.runtime.values.SmallIntArray;
 import i4gl.runtime.values.Varchar;
 
 @NodeInfo(shortName = "length")
@@ -37,38 +31,8 @@ public abstract class LengthBuiltinNode extends BuiltinNode {
     // Array length not actually supported by i4gl
 
     @Specialization
-    int length(CharArray array) {
+    int length(Array array) {
         return array.getSize();
-    }
-
-    @Specialization
-    int length(SmallIntArray array) {
-        return array.getSize();
-    }
-
-    @Specialization
-    int length(IntArray array) {
-        return array.getSize();
-    }
-
-    @Specialization
-    int length(BigIntArray array) {
-        return array.getSize();
-    }
-
-    @Specialization
-    int length(SmallFloatArray array) {
-        return array.getSize();
-    }
-
-    @Specialization
-    int length(FloatArray array) {
-        return array.getSize();
-    }
-
-    @Specialization
-    int length(Null string) {
-        return 0;
     }
 
     @Specialization